Moved test_db fixtures to conftest.py
